PEPPER-1498 . cleanup/refactor
